refactor(arch): huge architecture rework
added abstractions of platforms (OpenGL, Linux)
added abstractions and wrappers for renderer
updated ecs/graphics APIs when needed
Application and EntryPoint allow defining a client to the engine

+++ include/redoom/ecs/UpdateContext.hh
@@ -9,7 +9,7 @@
 namespace redoom::ecs
 {
 struct UpdateContext {
-  UpdateContext(long pelapsed_time,
+  UpdateContext(double pelapsed_time,
       ComponentManager& pcomponent_manager,
       EntityManager& pentity_manager) noexcept;
   UpdateContext(UpdateContext const&) noexcept = delete;
@@ -21,7 +21,7 @@ struct UpdateContext {
 
   void removeEntity(Entity entity) noexcept;
 
-  long elapsed_time;
+  double elapsed_time;
   ComponentManager& component_manager;
   EntityManager& entity_manager;
